refactor: reworked simulation class
Moved configuration out of the simulation class after reading about
dependency injection. The idea behind dependency injection is that your
dependencies are created and configured *outside* of the client that
uses them, so they can easily be swapped.

If a client is responsible for configuring its dependencies, then
either:

1) the client must know how to configure the dependency
2) the dependencies must provide a configuration method in the interface.

Both of these cause a coupling. With di, you decouple the dependencies
from the client so that they can easily be tested and swapped.

There is a library named [boost].di that can automate dependency
injection to reduce boilerplate (i.e. constructing all of the
dependencies with the correct arguments, etc). This may be useful in the
future.

Current idea for class structure

Simulation Builder
        Simulation
                HeatSolver
                Emitter
                        HeatSource
                        Waveform